--- src/main/java/nl/knaw/dans/catalog/api/TarPartOld.java
@@ -1,52 +1,0 @@
-/*
- * Copyright (C) 2022 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package nl.knaw.dans.catalog.api;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class TarPartOld {
-    @JsonProperty("part_name")
-    private String partName;
-    @JsonProperty("checksum_algorithm")
-    private String checksumAlgorithm;
-    @JsonProperty("checksum_value")
-    private String checksumValue;
-
-    public String getPartName() {
-        return partName;
-    }
-
-    public void setPartName(String partName) {
-        this.partName = partName;
-    }
-
-    public String getChecksumAlgorithm() {
-        return checksumAlgorithm;
-    }
-
-    public void setChecksumAlgorithm(String checksumAlgorithm) {
-        this.checksumAlgorithm = checksumAlgorithm;
-    }
-
-    public String getChecksumValue() {
-        return checksumValue;
-    }
-
-    public void setChecksumValue(String checksumValue) {
-        this.checksumValue = checksumValue;
-    }
-}
