+++ src/main/java/nl/knaw/dans/catalog/core/mapper/JsonMapper.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (C) 2022 DANS - Data Archiving and Networked Services (info@dans.knaw.nl)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package nl.knaw.dans.catalog.core.mapper;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class JsonMapper {
+    private static ObjectMapper objectMapper = new ObjectMapper();
+
+    public static String toString(Object object) {
+        try {
+            return objectMapper.writeValueAsString(object);
+        }
+        catch (JsonProcessingException e) {
+            return null;
+        }
+    }
+
+    public static Object toJson(String str) {
+        try {
+            return objectMapper.readTree(str);
+        }
+        catch (JsonProcessingException e) {
+            return null;
+        }
+    }
+}
