+++ src/main/java/nl/knaw/dans/catalog/core/mapper/TarMapper.java
@@ -13,20 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.catalog.core;
+package nl.knaw.dans.catalog.core.mapper;
 
-import nl.knaw.dans.catalog.db.TransferItem;
-import nl.knaw.dans.catalog.openapi.api.TransferItemDto;
+import nl.knaw.dans.catalog.db.Tar;
+import nl.knaw.dans.catalog.openapi.api.TarDto;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import org.mapstruct.factory.Mappers;
 
-@Mapper
-public interface TransferItemMapper {
-    TransferItemMapper INSTANCE = Mappers.getMapper(TransferItemMapper.class);
+@Mapper(uses = { TransferItemMapper.class })
+public interface TarMapper {
+    TarMapper INSTANCE = Mappers.getMapper(TarMapper.class);
 
-    @Mapping(source = "metadataJson", target = "metadata")
-    TransferItemDto transferItemToTransferItemDto(TransferItem transferItem);
-
+    TarDto tarToTarDto(Tar tar);
 }
